app/requirements.txt
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
PyMySQL==0.9.3
dialogflow==0.6.0
google-cloud-storage==1.20.0
clarifai==2.6.2
werkzeug==0.15.4
SpeechRecognition==3.8.1
google-cloud-texttospeech==0.5.0
